+++ index.js
@@ -20,6 +20,7 @@ const defaultOptions = {
   concurrency: 4,
   include: ["/"],
   exclude: [],
+  sitemap: false,
   userAgent: "ReactSnap",
   // 4 params below will be refactored to one: `puppeteer: {}`
   // https://github.com/stereobooster/react-snap/issues/120
@@ -646,6 +647,22 @@ const saveAsJpeg = ({ page, filePath, options, route }) => {
   return page.screenshot({ path: screenshotPath });
 };
 
+const buildSitemap = (routes, homepage) => {
+  const domain = homepage.replace(/\/$/, '')
+  return `
+    <?xml version="1.0" encoding="UTF-8"?>
+    <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+      ${routes.map(route => `
+        <url>
+          <loc>${domain + route}</loc>
+          <lastmod>${new Date().toISOString().split('T')[0]}</lastmod>
+          <priority>0.5</priority>
+        </url>
+      `).join(' ')}
+    </urlset>
+  `;
+};
+
 const run = async (userOptions, { fs } = { fs: nativeFs }) => {
   let options;
   try {
@@ -694,8 +711,9 @@ const run = async (userOptions, { fs } = { fs: nativeFs }) => {
   const basePath = `http://localhost:${options.port}`;
   const publicPath = options.publicPath;
   const ajaxCache = {};
-  const { http2PushManifest } = options;
+  const { http2PushManifest, sitemap } = options;
   const http2PushManifestItems = {};
+  const sitemapItems = []
 
   await crawl({
     options,
@@ -708,6 +726,9 @@ const run = async (userOptions, { fs } = { fs: nativeFs }) => {
         cacheAjaxRequests,
         preconnectThirdParty
       } = options;
+
+      if (!route.endsWith("/404.html")) sitemapItems.push(route)
+
       if (
         preloadImages ||
         cacheAjaxRequests ||
@@ -876,6 +897,21 @@ const run = async (userOptions, { fs } = { fs: nativeFs }) => {
           `${destinationDir}/http2-push-manifest.json`,
           JSON.stringify(manifest)
         );
+      }
+
+      if (sitemap) {
+        if (!options.homepage) {
+          console.log('⚠️   To generate a sitemap.xml a domain is required, add homepage to package.json');
+          return;
+        }
+
+        const xml = buildSitemap(sitemapItems, options.homepage);
+
+        fs.writeFileSync(
+          `${destinationDir}/sitemap.xml`,
+          xml.replace(/^\s+/gm, '')
+        );
+        console.log('Sitemap generated!');
       }
     }
   });